entry styling
