+++ lotti/lib/widgets/pages/journal_page2.dart
@@ -1,19 +1,18 @@
-import 'dart:io';
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:lotti/blocs/audio/player_cubit.dart';
 import 'package:lotti/blocs/journal/persistence_cubit.dart';
 import 'package:lotti/blocs/journal/persistence_state.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/widgets/journal/entry_modal_widget2.dart';
-import 'package:lotti/widgets/journal/entry_tools.dart';
-import 'package:lotti/widgets/journal/journal_list_item.dart';
+import 'package:lotti/widgets/journal/journal_card.dart';
 
 class JournalPage2 extends StatefulWidget {
-  const JournalPage2({this.navigatorKey, required this.child});
+  const JournalPage2({
+    Key? key,
+    this.navigatorKey,
+    required this.child,
+  });
 
   final Widget child;
   final GlobalKey? navigatorKey;
@@ -23,13 +22,8 @@ class JournalPage2 extends StatefulWidget {
 }
 
 class _JournalPage2State extends State<JournalPage2> {
-  late TextEditingController _textEditingController;
-
-  int _currentRoute = 0;
-
   @override
   void initState() {
-    _textEditingController = TextEditingController();
     super.initState();
   }
 
@@ -66,35 +60,7 @@ class _JournalPage2State extends State<JournalPage2> {
                             entries.length,
                             (int index) {
                               JournalEntity item = entries.elementAt(index);
-
-                              return Card(
-                                child: ListTile(
-                                  leading: const FlutterLogo(),
-                                  //title: Text('Item ${index + 1}'),
-                                  title: JournalListItem2(item: item),
-                                  enabled: true,
-                                  onTap: () {
-                                    item.mapOrNull(
-                                        journalAudio: (JournalAudio audioNote) {
-                                      context
-                                          .read<AudioPlayerCubit>()
-                                          .setAudioNote(audioNote);
-                                    });
-
-                                    _currentRoute = index;
-                                    Navigator.of(context).push(
-                                      MaterialPageRoute(
-                                        builder: (BuildContext context) {
-                                          return DetailRoute(
-                                            item: item,
-                                            index: index,
-                                          );
-                                        },
-                                      ),
-                                    );
-                                  },
-                                ),
-                              );
+                              return JournalCard(item: item, index: index);
                             },
                           ),
                         );
@@ -107,34 +73,6 @@ class _JournalPage2State extends State<JournalPage2> {
           },
         );
       },
-    );
-  }
-}
-
-class DetailRoute extends StatelessWidget {
-  DetailRoute({
-    Key? key,
-    required this.item,
-    required this.index,
-  }) : super(key: key);
-
-  final int index;
-  JournalEntity item;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          df.format(item.meta.dateFrom),
-          style: TextStyle(color: AppColors.entryBgColor),
-        ),
-        backgroundColor: AppColors.headerBgColor,
-      ),
-      body: EntryModalWidget2(
-        item: item,
-        docDir: Directory('docDir'),
-      ),
     );
   }
 }
